Accumulate lint results and call `fill_panel` once
Previously, we just limited redraws of the panel, see 2cd32b7

We now effectively throttle calls to `fill_panel`. The actual
throttler uses the same technique as before, t.i. bouncing to the
worker thread.
The main difference is that we *accumulate* `LINT_RESULT` events for
the timespan of one tick on the worker thread, then decide how to
proceed on the cumulated data, and eventually `fill_panel` once.

Noteable: If we see a result for the current active view, we draw
immediately. (This solves a drawback of the previous solution.)